requirements.txt
pandas
packaging
gym
numpy
dill
matplotlib
torch
tqdm
gitpython
scikit_learn
tensorboardX
tables
mujoco-py==1.50.1.68
